requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
python-multipart==0.0.20
tensorflow==2.19.0
pillow==11.3.0
numpy<2.0
